added bannerModal in config settings with new banner style
